[Refactor] Extract GasEstimation to separate class, move models to separate files for TransactionConfigurator #4559
