+++ AlphaWallet/Gas/Models/EtherscanGasPriceEstimator.swift
@@ -37,9 +37,7 @@ fileprivate extension RPCServer {
         case .main, .binance_smart_chain, .heco, .polygon:
             return etherscanApiRoot?.appendingQueryString("\("module=gastracker&action=gasoracle")\(apiKeyParameter)")
         case .artis_sigma1, .artis_tau1, .binance_smart_chain_testnet, .callisto, .poa, .sokol, .classic, .xDai, .heco_testnet, .fantom, .fantom_testnet, .avalanche, .avalanche_testnet, .mumbai_testnet, .cronosTestnet, .custom, .arbitrum, .arbitrumRinkeby, .kovan, .ropsten, .rinkeby, .goerli, .optimistic, .optimisticKovan, .palm, .palmTestnet, .klaytnCypress, .klaytnBaobabTestnet:
-            return nil
-        case .klaytnCypress: return nil
-        case .klaytnBaobabTestnet: return nil
+            return nil 
         }
     }
 }
